test(coverage): project-export-import.ts Fix-and-Close — 67.24% → 76.46% lines (isolated)
Consolidated project-export-import-v4.test.ts into project-export-import.test.ts.
Both files mocked the same modules (../lib/prisma, @shogo/shared-runtime,
@shogo-ai/sdk/agent, ../lib/runtime-token) with competing factories — when run
in the same bun-test process the loser's tests saw the wrong mock state and
failed (password-protected round-trip + workspace-file-on-disk, 2 fails).

Merged via one shared mock harness with toggle flags:
- bundleResponse / bundleThrows → drive AgentClient.getWorkspaceBundle()
- s3UploadResult / s3SyncReturnsNull → drive createS3SyncForProject()
- per-test process.env.KUBERNETES_SERVICE_HOST / S3_WORKSPACES_BUCKET toggles

beforeEach now resets toggles + env to local mode so k8s tests opt in
explicitly. knative-project-manager mock added at module scope (lazy-imported
only when k8s branch fires).

Before: 29 pass / 2 fail (cross-file mock pollution)
After:  31 pass / 0 fail (Fix-and-Close achieved)
Isolated coverage on project-export-import.ts: 67.24% -> 76.46% lines (+9.22pp)
                                               73.55% -> 77.42% funcs (+3.87pp)
